requirements.txt
python-docx>=0.8.11
pandas>=1.3.0
openpyxl>=2.6.0
